eeg_depression_detection/requirements.txt
# Core Deep Learning
torch>=2.0
torch-geometric>=2.3
torchvision>=0.15

# Scientific Computing
numpy>=1.24
scipy>=1.10
pandas>=2.0

# EEG Processing
mne>=1.5
pywavelets>=1.4

# Visualization
matplotlib>=3.7
seaborn>=0.12
plotly>=5.14

# Explainability (XAI)
captum>=0.6  # Integrated Gradients, LRP
shap>=0.42   # For comparison with baseline

# Machine Learning Utilities
scikit-learn>=1.3
imbalanced-learn>=0.11

# Utilities
tqdm>=4.65
pyyaml>=6.0
python-dotenv>=1.0
h5py>=3.9

# Testing
pytest>=7.3
pytest-cov>=4.1

# Jupyter
jupyter>=1.0
ipywidgets>=8.0
